Refactoring and editing pictures
